[text_service]Prevent pylint to check signatures
When having a base class and a derived class, pylint can complain
because some methods have the same signatures. In particular, this
happens if the function signature spans more than 4 rows, triggering
the similarity checker. This is the case for the function
lemmas_frequency.
I disable this pylint feature.

For more info: https://github.com/PyCQA/pylint/issues/3619